restruct 0.3 based 1.6
完成楼负责人操作分包
getdata.deliverer_get_community -> staff.deliverer.get_community
getdata.deliverer_get_building -> staff.deliverer.get_building
getdata.get_delivery_list -> staff.deliverer.get_delivery_list
修改相应url
